Ensure every Page holds a scoped URL, even if it's target URL isn't scoped.
iOS is very weird in that it treats certain URLs as security
capabilities, you have to present the correct URL to unlock a given
resource, and you can't just unlock all the URLs at discovery time and
keep them unlocked (apparently this leaks kernel resources?!). So we
have to always carry around a separate URL just for unlocking the file
that we will later access.