Fix source view changes not showing up in the WebKit preview correctly.
This was caused by our save throttling combined with the recent fixes to sub resource loading. Since WebKit does not allow us to load a string from disk while still pulling resources from the filesystem, we have to work around that by making JavaScript change the page contents out while the file is still being saved. Once we save the file we force a reload from disk, as we currently do not implement HTML parsing to catch changes to the `<html>` element and we don't want WebKit to be missing data in *any* circumstance.